Add the prototype for srhw_sric_enumerate_dev_class
This is a new function, as discussed on the list, which wraps sric_enumerate_devices
but filters the results for devices only of the specified class. It also expects an srhw_ctx
structure, which hasn't been defined yet, as the whole library is moving to a context based thread-safing
approach.