Add the struct srhw_motor_ctx
This struct holds all the info relating to motors that needs to be stored within
an srhw_ctx structure, which hasn't been fully defined yet. (Should happen soon!)

This struct is pretty much a copy-paste of the two static global variables in motors.c
which are about to be removed in the name of thread-safing.