new file srhwctx.h - Contains the declaration of the srhw_ctx struct
This is the structure containing all per-context info. It simply contains all
the *_ctx structures declared by other parts of the library (eg: motor.h's srhw_motor_ctx)
in order to ensure modularity is maintained. (This way, we don't have to declare stuff related to
motors outside of motor.h, for example)