Remove #include sricdriv.h as this file no longer exists.
Furthermore, its replacement does not need to be included in srhw.h either.
This is because the old version required initialisation, and so treating it as a driver
was the logical way to avoid extra initialisation code. The new system, however, requires
that we export a srhw_context_get routine, which doesn't exist yet.