Improve makefile, fix warnings.
Improve the makefile (ensure that the correct directories exist and that
sricd is built). Add -Wall -Werror -Wextra -pedantic and make it
compile (also fix what looks like a logic bug in the process). It still
doesn't link, because for some reason it's trying to build an
executable rather than a library.